UPDATE: 기본과제 2차 완성
