+++ src/basic/observable/index.js
@@ -1,0 +1,15 @@
+class Observable {
+  constructor() {
+    this.observers = [];
+  }
+
+  subscribe(observer) {
+    this.observers.push(observer);
+  }
+
+  notify(data) {
+    this.observers.forEach((observer) => observer(data));
+  }
+}
+
+export const lastSelObservable = new Observable();
