+++ src/basic/services/product.js
@@ -1,0 +1,5 @@
+import { prodList } from '../constant';
+
+export const getProductById = (productId) => {
+  return prodList.find((p) => p.id === productId);
+};
